+++ internal/libinjection/libinjection_stub.go
@@ -1,0 +1,7 @@
+//go:build !libinjection_cgo
+
+package libinjection
+
+func Register() {
+	// we don't do anything
+}