+++ internal/re2/re2.go
@@ -1,0 +1,9 @@
+//go:build re2_cgo
+
+package re2
+
+import "github.com/corazawaf/coraza-wasilibs"
+
+func Register() {
+	wasilibs.RegisterRX()
+}