+++ internal/re2/re_stub.go
@@ -1,0 +1,7 @@
+//go:build !re2_cgo
+
+package re2
+
+func Register() {
+	// we don't do anything
+}